Created module for set status
